+++ tparf.Api/Entities/CartItem.cs
@@ -2,9 +2,9 @@
 {
     public class CartItem
     {
-        public int Id { get; set; }
-        public int CartId { get; set; }
-        public int ProductId { get; set; }
+        public long Id { get; set; }
+        public long CartId { get; set; }
+        public long ProductId { get; set; }
         public int Qty { get; set; }
     }
 }
