+++ tparf.Api/Entities/User.cs
@@ -6,12 +6,12 @@ namespace tparf.Api.Entities
     public class User
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string UserName { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
         public string? OrganizationName { get; set; }
-        public int RoleId { get; set; }
+        public long RoleId { get; set; }
         [ForeignKey("RoleId")]
         public Role Role { get; set; }
