+++ tparf.Api/Repositories/Contracts/IShoppingCartRepository.cs
@@ -6,9 +6,9 @@ namespace tparf.Api.Repositories.Contracts
     public interface IShoppingCartRepository
     {
         Task<CartItem> AddItem(CartItemToAddDto cartItemToAddDto);
-        Task<CartItem> UpdateQty(int id, CartItemQtyUpdateDto cartItemQtyUpdateDto);
-        Task<CartItem> DeleteItem(int id);
-        Task<CartItem> GetItem(int id);
-        Task<IEnumerable<CartItem>> GetItems(int userId);
+        Task<CartItem> UpdateQty(long id, CartItemQtyUpdateDto cartItemQtyUpdateDto);
+        Task<CartItem> DeleteItem(long id);
+        Task<CartItem> GetItem(long id);
+        Task<IEnumerable<CartItem>> GetItems(long userId);
     }
 }
