+++ tparf.Models/Dtos/CartItemQtyUpdateDto.cs
@@ -8,7 +8,7 @@ namespace tparf.Models.Dtos
 {
     public class CartItemQtyUpdateDto
     {
-        public int CartItemId { get; set; }
+        public long CartItemId { get; set; }
         public int Qty { get; set;}
     }
 }
