+++ tparf.Web/Pages/ProductsBase.cs
@@ -48,14 +48,14 @@ namespace tparf.Web.Pages
 
         }
 
-        protected IOrderedEnumerable<IGrouping<int, ProductDto>> GetGroupedProductsByCategory()
+        protected IOrderedEnumerable<IGrouping<long, ProductDto>> GetGroupedProductsByCategory()
         {
             return from product in Products
                    group product by product.CategoryId into prodByCatGroup
                    orderby prodByCatGroup.Key
                    select prodByCatGroup;
         }
-        protected string GetCategoryName(IGrouping<int, ProductDto> groupedProductDtos)
+        protected string GetCategoryName(IGrouping<long, ProductDto> groupedProductDtos)
         {
             return groupedProductDtos.FirstOrDefault(pg => pg.CategoryId == groupedProductDtos.Key).CategoryName;
         }
