+++ tparf.Web/Services/Contracts/IShoppingCartService.cs
@@ -4,9 +4,9 @@ namespace tparf.Web.Services.Contracts
 {
     public interface IShoppingCartService
     {
-        Task<List<CartItemDto>> GetItems(int userId);
+        Task<List<CartItemDto>> GetItems(long userId);
         Task<CartItemDto> AddItem(CartItemToAddDto cartItemToAddDto);
-        Task<CartItemDto> DeleteItem(int id);
+        Task<CartItemDto> DeleteItem(long id);
         Task<CartItemDto> UpdateQty(CartItemQtyUpdateDto cartItemQtyUpdateDto);
 
         event Action<int> OnShoppingCartChanged;
